move pod_inherited.py to pod.py
